+++ src/Controller/RegisterController.php
@@ -11,13 +11,15 @@ use Symfony\Component\Routing\Annotation\Route;
 use App\Entity\User;
 use Symfony\Component\Security\Core\Encoder\UserPasswordEncoderInterface;
 use App\Repository\LeagueRepository;
+use App\Repository\UserRepository;
+use Symfony\Component\Form\FormError;
 
 class RegisterController extends AbstractController
 {
     /**
      * @Route("/signup", name="app_signup")
      */
-    public function register(Request $request, UserPasswordEncoderInterface $passwordEncoder, LeagueRepository $leagueRepository)
+    public function register(Request $request, UserPasswordEncoderInterface $passwordEncoder, LeagueRepository $leagueRepository, UserRepository $userRepository)
     {
         $form = $this->createFormBuilder()
             ->add('firstName')
@@ -41,14 +43,49 @@ class RegisterController extends AbstractController
 
         if ($form->isSubmitted()) {
             $data = $form->getData();
+            // checks if email exists
+            if ($userRepository->findOneBy(['email' => $data['email']])) {
+                $error = new FormError('Email Already Exists');
+                $form->get('email')->addError($error);
+                return $this->render(
+                    'register/index.html.twig',
+                    [
+                        'form' => $form->createView()
+                    ]
+                );
+            }
+            // checks if username exists
+            if ($userRepository->findOneBy(['username' => $data['username']])) {
+                $error = new FormError('Username Already Exists');
+                $form->get('username')->addError($error);
+                return $this->render(
+                    'register/index.html.twig',
+                    [
+                        'form' => $form->createView()
+                    ]
+                );
+            }
+            // gets league from league table based on ref code given
+            $league = $leagueRepository->findOneBy(['refCode' => $data['refCode']]);
+            // checks if league exists
+            if (!$league) {
+                $error = new FormError('Invalid Ref Code');
+                $form->get('refCode')->addError($error);
+                return $this->render(
+                    'register/index.html.twig',
+                    [
+                        'form' => $form->createView()
+                    ]
+                );
+            }
             $user = new User();
             $user->setFirstName($data['firstName']);
             $user->setLastName($data['lastName']);
             $user->setUsername($data['username']);
             $user->setEmail($data['email']);
-            // get league name from league table
-            $league = $leagueRepository->findOneBy(['refCode' => $data['refCode']]);
-            $user->setLeagueID($league->getId());
+
+            $user->setLeagueName($league->getLeagueName());
+            $user->setLeague($league);
             $user->setPassword(
                 $passwordEncoder->encodePassword($user, $data['password'])
             );