+++ src/Entity/Stats.php
@@ -1,0 +1,120 @@
+<?php
+
+namespace App\Entity;
+
+use Doctrine\ORM\Mapping as ORM;
+
+/**
+ * @ORM\Entity(repositoryClass="App\Repository\StatsRepository")
+ */
+class Stats
+{
+    /**
+     * @ORM\Id()
+     * @ORM\GeneratedValue()
+     * @ORM\Column(type="integer")
+     */
+    private $id;
+
+    /**
+      * @ORM\ManyToOne(targetEntity="User", inversedBy="id")
+      */
+    private $player;
+
+    /**
+     * @ORM\Column(type="string", length=255)
+     */
+    private $playerHit;
+
+    /**
+     * @ORM\Column(type="date")
+     */
+    private $date;
+
+    /**
+     * @ORM\Column(type="string", length=255)
+     */
+    private $type;
+
+    /**
+     * @ORM\Column(type="float")
+     */
+    private $points;
+
+    public function getId(): ?int
+    {
+        return $this->id;
+    }
+
+    public function getPlayerHit(): ?string
+    {
+        return $this->playerHit;
+    }
+
+    public function setPlayerHit(string $playerHit): self
+    {
+        $this->playerHit = $playerHit;
+
+        return $this;
+    }
+
+    public function getType(): ?string
+    {
+        return $this->type;
+    }
+
+    public function setType(string $type): self
+    {
+        $this->type = $type;
+
+        return $this;
+    }
+
+    public function getPoints(): ?float
+    {
+        return $this->points;
+    }
+
+    public function setPoints(float $points): self
+    {
+        $this->points = $points;
+
+        return $this;
+    }
+
+    public function getStats(): ?User
+    {
+        return $this->stats;
+    }
+
+    public function setStats(?User $stats): self
+    {
+        $this->stats = $stats;
+
+        return $this;
+    }
+
+    public function getPlayer(): ?User
+    {
+        return $this->player;
+    }
+
+    public function setPlayer(?User $player): self
+    {
+        $this->player = $player;
+
+        return $this;
+    }
+
+    public function getDate(): ?\DateTimeInterface
+    {
+        return $this->date;
+    }
+
+    public function setDate(\DateTimeInterface $date): self
+    {
+        $this->date = $date;
+
+        return $this;
+    }
+}